functions/requirements.txt
python-telegram-bot>=21.0
firebase-admin==6.4.0
firebase-functions==0.1.0
requests==2.31.0
pytz==2024.1
APScheduler==3.10.4
aiohttp==3.9.3
python-dotenv==1.0.1
